# docs/api/openapi/process_heat_openapi.yaml
openapi: 3.0.3
info:
  title: GreenLang Process Heat API
  description: |
    The GreenLang Process Heat API provides programmatic access to industrial
    process heat optimization, emissions monitoring, and regulatory compliance
    capabilities.

    ## Features
    - Process heat workflow orchestration
    - Real-time emissions calculations (Scope 1, 2, 3)
    - Regulatory compliance reporting (CSRD, CBAM, EUDR)
    - Complete audit trail with SHA-256 provenance

    ## Authentication
    All endpoints require OAuth 2.0 Bearer token authentication.

    ## Rate Limiting
    - Standard: 100 requests/minute
    - Professional: 300 requests/minute
    - Enterprise: 1000 requests/minute
  version: 1.0.0
  contact:
    name: GreenLang API Support
    email: support@greenlang.io
    url: https://docs.greenlang.io/process-heat
  license:
    name: Proprietary
    url: https://greenlang.io/terms
  x-logo:
    url: https://greenlang.io/logo.png
    altText: GreenLang Logo
servers:
  - url: https://api.greenlang.io/v1
    description: Production server
  - url: https://staging-api.greenlang.io/v1
    description: Staging server
  - url: https://dev-api.greenlang.io/v1
    description: Development server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /agents:
    get:
      tags:
        - Agents
      summary: List registered agents
      description: Retrieve list of all registered agents
      operationId: listAgents
      responses:
        '200':
          description: Agents listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsListResponse'
    post:
      tags:
        - Agents
      summary: Register agent
      description: Register a new agent with the orchestrator
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegistrationResponse'
  /agents/{agent_id}:
    get:
      tags:
        - Agents
      summary: Get agent status
      description: Retrieve status of a specific agent
      operationId: getAgentStatus
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatusResponse'
    delete:
      tags:
        - Agents
      summary: Deregister agent
      description: Remove an agent from the orchestrator
      operationId: deregisterAgent
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent deregistered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================================
  # Safety APIs
  # ============================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 Bearer token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server integrations

  # ============================================================================
  # Parameters
  # ============================================================================
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
        timestamp:
          type: string
          format: date-time
...
    ResponseMetadata:
      type: object
      properties:
        request_id:
          type: string
        processing_time_ms:
          type: number
...
    AgentsListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                agents:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentStatus'
    AgentRegistrationRequest:
      type: object
      required:
        - agent_id
        - agent_type
        - name
      properties:
        agent_id:
          type: string
        agent_type:
          type: string
        name:
          type: string
        capabilities:
          type: array
          items:
            type: string
        endpoint:
          type: string
          format: uri
    AgentRegistrationResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                agent_id:
                  type: string
                registered:
                  type: boolean
                registered_at:
                  type: string
                  format: date-time
    AgentStatus:
      type: object
      properties:
        agent_id:
          type: string
        agent_type:
          type: string
        name:
          type: string
        health:
          type: string
          enum: [healthy, degraded, unhealthy, offline, maintenance]
        version:
          type: string
        last_heartbeat:
          type: string
          format: date-time
        active_tasks:
          type: integer
        capabilities:
          type: array
          items:
            type: string
    AgentStatusResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AgentStatus'

    # --------------------------------------------------------------------------
    # Safety Schemas
    # --------------------------------------------------------------------------
...